+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ## [1.3.0] Unreleased
 
 * Make HRID optional in Instance form (UIIN-304)
-* Validate HRID for uniqueness in Instance form UIIN-288
+* Validate HRID for uniqueness in Instance form (UIIN-288)
 * Update `stripes-form` dependency to v1.0.0
 * Remove cataloging level (UIIN-312)
 * Remove API dependency on `cataloging-levels` (UIIN-312)
 * More stable inventory-search tests. Refs UIIN-306.
+* Remove `instance.edition` add `instance.editions` (UIIN-299)
 
 ## [1.2.3 (hot fix)](https://github.com/folio-org/ui-inventory/tree/v1.2.3) (2018-09-19)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.2.2...v1.2.3)